requirements.txt
streamlit==1.32.2

langchain==0.2.17
langchain-core==0.2.43
langchain-community==0.2.19
langchain-text-splitters==0.2.4
langchain-groq==0.1.9
langchain-huggingface==0.0.3

chromadb==0.5.5

sentence-transformers==2.7.0
transformers==4.41.2
torch==2.2.2

pypdf
python-dotenv

pysqlite3-binary; platform_system != "Windows"
